@@ -7,14 +7,15 @@
     "@fortawesome/free-solid-svg-icons": "^6.4.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
     "bootstrap": "^5.2.3",
     "i18next": "^22.4.13",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.7.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^12.3.1",
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
