@@ -1,12 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from "react-router-dom"
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
 import { App } from './App';
 import { AlertProvider } from './components/Alert/AlertContext';
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <BrowserRouter>
       <AlertProvider>
@@ -14,5 +14,4 @@ ReactDOM.render(
       </AlertProvider>
     </BrowserRouter>
   </React.StrictMode>,
-  document.getElementById('root')
 );
